fix(widget): do not request wallet capabilities for wallet-connect
