+++ libs/wallet/src/index.ts
@@ -8,6 +8,7 @@ export * from './constants'
 // Hooks
 export * from './api/hooks'
 export { useWalletCapabilities } from './api/hooks/useWalletCapabilities'
+export { useWidgetProviderMetaInfo } from './api/hooks/useWidgetProviderMetaInfo'
 export { useSendBatchTransactions } from './api/hooks/useSendBatchTransactions'
 export type { SendBatchTxCallback } from './api/hooks/useSendBatchTransactions'
 export * from './web3-react/hooks/useWalletMetadata'
